+++ src/data/number/formatNum.js
@@ -1,0 +1,18 @@
+import isSymbol from '../type/isSymbol'
+
+/**
+ * 格式化数字
+ * @param {string | number} originNum 数据
+ * @returns {number}
+ */
+ function formatNum(value) {
+  if (typeof value === 'number') {
+    return value
+  }
+  if (isSymbol(value)) {
+    return Number.NaN
+  }
+  return Number(value)
+}
+
+export default formatNum