+++ history.md
@@ -5,6 +5,7 @@
 - updateList中option.format不返回需要push的数据，因为此时需要的对象都是复杂格式，内存指针形式，因此format仅对对象数据做格式化，返回值为是否添加到数组中
 - updateDataWidthOption实现total模式
 - updateDataWidthOption参数顺序调整
+- 添加isSymbol/formatNum，优化getType的symbol判断
 
 ### 2.0.18
 - current注释